fixed email error where apple users were getting app store prefixes in links. added auth protection to controllers and took away ability to edit other users. added features to admin protect certain parts of the site. added link to songs

+++ resources/views/user/users/index.blade.php
@@ -8,29 +8,34 @@
         <table class="table table-bordered table-striped table-hover">
             <thead>
                 <tr>
-                    <th> Name </th><th> Bio </th><th>Actions</th>
+                    <th> Name </th><th> Bio </th>
+                    @if(Auth::user()->hasRole('admin'))
+                        <th>Actions</th>
+                    @endif
                 </tr>
             </thead>
             <tbody>
             @foreach($users as $item)
                 <tr>
                     <td>{{ $item->name }}</td><td>{{ $item->bio }}</td>
-                    <td>
-                        <a href="{{ url('/users/' . $item->id) }}" class="btn btn-success btn-xs" title="View user"><span class="glyphicon glyphicon-eye-open" aria-hidden="true"/></a>
-                        <a href="{{ url('/users/' . $item->id . '/edit') }}" class="btn btn-primary btn-xs" title="Edit user"><span class="glyphicon glyphicon-pencil" aria-hidden="true"/></a>
-                        {!! Form::open([
-                            'method'=>'DELETE',
-                            'url' => ['/users', $item->id],
-                            'style' => 'display:inline'
-                        ]) !!}
-                            {!! Form::button('<span class="glyphicon glyphicon-trash" aria-hidden="true" title="Delete user" />', array(
-                                    'type' => 'submit',
-                                    'class' => 'btn btn-danger btn-xs',
-                                    'title' => 'Delete user',
-                                    'onclick'=>'return confirm("Confirm delete?")'
-                            ));!!}
-                        {!! Form::close() !!}
-                    </td>
+                    @if(Auth::user()->hasRole('admin'))
+                        <td>
+                            <a href="{{ url('/users/' . $item->id) }}" class="btn btn-success btn-xs" title="View user"><span class="glyphicon glyphicon-eye-open" aria-hidden="true"/></a>
+                            <a href="{{ url('/users/' . $item->id . '/edit') }}" class="btn btn-primary btn-xs" title="Edit user"><span class="glyphicon glyphicon-pencil" aria-hidden="true"/></a>
+                            {!! Form::open([
+                                'method'=>'DELETE',
+                                'url' => ['/users', $item->id],
+                                'style' => 'display:inline'
+                            ]) !!}
+                                {!! Form::button('<span class="glyphicon glyphicon-trash" aria-hidden="true" title="Delete user" />', array(
+                                        'type' => 'submit',
+                                        'class' => 'btn btn-danger btn-xs',
+                                        'title' => 'Delete user',
+                                        'onclick'=>'return confirm("Confirm delete?")'
+                                ));!!}
+                            {!! Form::close() !!}
+                        </td>
+                    @endif
                 </tr>
             @endforeach
             </tbody>
